pyproject.toml
[project]
name = "authfilter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiomysql>=0.2.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.16.4",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "greenlet>=3.2.3",
    "htmx>=0.0.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pymysql>=1.1.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-httpx>=0.35.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]

[project.scripts]
build-image = "scripts.build_image:main"
push-image = "scripts.push_image:main"
build-and-push-image = "scripts.build_and_push_image:main"
retag-latest = "scripts.retag_latest:main"
dev = "uvicorn:run"
test = "pytest:main"
migrate = "alembic:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
